refactor: 파일명 변경 및 @types 추가
