+++ @types/opyional.ts
@@ -1,0 +1,5 @@
+export interface IOptional<T> {
+    isPresent(): boolean
+
+    get(): T
+}
