+++ utils/stringUtils.ts
